Create Write_a_function.py
It was very tricy at first, they gave:
leap=False.
which was very confusing for a beginner like me, I was using function as the main result. It was unnecessary to write leap = False. when the boolean function can itself find out it is True or False. I thought I need to add " leap = True."